pyproject.toml
[project]
name = "mvp-ticketing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.23",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.3",
    "gunicorn>=23.0.0",
    "whitenoise[brotli]>=6.9.0",
]
